Standardize vector registers
The way that vector registeres are defined in DWARF is as if they are
arrays but with a special attribute DW_AT_GNU_vector. Although this
has been standard practice for a long time and that attribute has been
used by all compilers, this practice has been overlooked in the DWARF
standardization process. This adds the modifications to the DWARF
standard to codify the current practice.

Since subsequent changes text in 001 refers to places that need to be
changed, it to avoid textual conflicts, this standardization of the
vector registers was inserted at the beginning of the stack of DWARF
issues. If people are not happy with that, then I'm more than willing
to submit that into the DWARF issue queue indpendently but some of the
text in the clarifications eval document would still need to be
updated.

Signed-off-by: Ben Woodard <woodard@redhat.com>