Further specify vector registers
Both ARM and RISC-V have variable width vector registers which are
hardware implementation defined. The original formulation of this
issue didn't provide a way to specify those. This version adds a new
attribute which allows a subrange to specify a variable vector
register subrange.

This also moves the requirements for vector register array types down
to their definition in section 5.5 rather than repeating the text for
each DWARF operator.

Signed-off-by: Ben Woodard <woodard@redhat.com>