Generalize creation of undefined location descriptions
DWARF 5 uses an empty expression to indicate an undefined location
description. Since [Allow location description on the DWARF evaluation
stack] allows location descriptions to be created on the stack, it is
necessary to have an explicit way to specify an undefined location
description.

For example, the DW_OP_select_bit_piece operation proposed in
[Operations to Create Vector Composite Location Descriptions] takes more
than one location description on the stack. Without this ability, it is
not possible to specify that a particular one of the input location
descriptions is undefined.