Generalize creating composite location descriptions
In DWARF 5, a composite location description is created by one or more
DW_OP_*piece operations. The end of the composite creation is indicated
by the end of the operation expression. This means a single operation
expression can only create a single composite location description, and
it must be created as the last step of the expression.

After [Allow location description on the DWARF evaluation stack]
proposal, a way is required to allow, possibly multiple, composite
location descriptions to be created anywhere in an operation expression.
This allows general composition of operations that create location
descriptions, and allows operations to be defined that take multiple
location descriptions (for example the [Operations to Create Vector
Composite Location Descriptions] proposal).

To support this, an explicit operation that indicates the end of the
definition of a composite location description is required. This can
also be implied if the end of a DWARF expression is reached, allowing
DWARF 5 expressions to remain legal.