Move clarifications to top of the queue
The address space proposal seems to depend on concepts that are
introduced in two editorial clarifications which were originally going
to be introudced before the original locdesc proposal. The locdesc
proposal was rewritten so that it no longer depends on those
proposals. However, it appears that the address space proposal still
relies on them. They appear to be rather self contained and ready to
apply. They should either be introduced on their own or folded into
the address space proposal.

Signed-off-by: Ben Woodard <woodard@redhat.com>