Update Overlay proposal addressing fixmes
The primary goal of this update was to address the FIXME's.

The first one was a request to explain how overlays could be used to
concatenate values. This was covered in a November 24th email by
Cary. The second fixme was to rewrite the definition of overlay not in
terms of how the current piece operators work. The original intent of
defining the overlay operators in terms of the piece operators was to
help current developers writing either producers or consumers to
understand the new overlay operators. To a certain extent this weakens
the justification for the overlay operators. This update seeks to help
developers by providing more examples and building the case for
overlay operators. At that point, it seems like the text that defines
the overlay operators in terms of the old piece operators is no longer
needed. Currently, the text with those examples is found in the
opening discussion but it could be moved or duplicated into Appendix D
and or E. So that it ends in the specification document.

Signed-off-by: Ben Woodard <woodard@redhat.com>