Respond to Cary's comments; fix diagrams
* Respond to Cary's comments

Cary reviewed my version of the Overlay proposal and made some
requested changes. He did not like the way that I did the ASCII art
diagram. I redid them to try to address his concerns.

Changes:
  * The diagrams are all redone.
  * The Several bugs in the examples were fixed
  * Instead of appending them overlay operators to section 3.11 create
    a new section 3.12

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Fix rendering of some diagrams

Some of the markdown was wrong. Fix it.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* untabify

It seems like emacs was using tabs rather than spaces which was
messing up the rendering of the markdown with the ASCII
art. Untabifying should fix many problems.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Fix a typo

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Address Ron's concerns

Ron Brender sent me an email asking for some changes in how the
operators are specified in the text that will be included in the
standard. I changed the descriptions of the operators to match the
template that he asked for.

Signed-off-by: Ben Woodard <woodard@redhat.com>

---------

Signed-off-by: Ben Woodard <woodard@redhat.com>