Fix SIMD Lane example.
Markus pointed out that C/C++'s decay to a pointer semantics breaks my
example. However, Cary had a clever workaround that turned the pointer
into an implicit pointer which arguably it really is when some of its
values now exist in vector registers.

Bonus: a couple tiny touch-ups to text. One was a formatting error.

Signed-off-by: Ben Woodard <woodard@redhat.com>