Address Baris's observation about length
Address Baris's concern that I was misleading readers in my
introductory examples about the length of composite storage when it is
placed over undefined storage. He was correct, the text seemed to
indicate that there was a finite length to the composite storage that
was created. Tried to fix the text.

Signed-off-by: Ben Woodard <woodard@redhat.com>