Rework examples (#107)
* Rework examples to use overlay

Since the desire is to use overlays rather than piece and bit piece,
this patch reworks the existing examples to use overlays rather than
piece and bit piece.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Fix formatting

Fix the formatting to better align with other proposals.
Fix a couple of minor mistakes.
Note a particular point of uncertainty that existed in the original example.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Fixed a couple of overlay examples

Double checking my work, I found a couple of errors and fixed them.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Rework SIMD example

The SIMD example didn't include a way to show the value of elements in
the arrays which are at partly stored in memory and partly in vector
registers.  Add overlays to the example to show how they can express
this common scenario.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Fix formatting in SIMD example

Fix a minor formatting problem in the SIMD example.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Fix more formatting

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Fix more formatting

More formatting problems that I didn't catch the first time.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Try to address Tony's concerns.

Tony had several concerns with the overlay proposal as I had
originally written it. This patch tries to address several of those
concerns.

* He wanted it to be clear that size of an overlay is the size of the
  underlying physical storage. This is distinctly different than the
  size of the composite created using the piece operators which has a
  fixed size. I had assumed this and it was implicit in what I wrote
  but I never explictly stated that. Since readers are likely familiar
  with how the piece operator works, they are likely to use their
  understanding of those when trying to understand overlays. This can
  lead to some confusion.

* He also wanted me to define that accessing an offset beyond the
  length of the defined portion of location causes the size of
  location to grow.

* He wanted it explicitly stated that the resulting location is a
  tuple and thus has a default offset of zero.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Address many of Cary's comments.

Cary made several comments to the git PR. This addresses several of
them but I didn't get through all of them. The remaining comments will
be addressed in a subsequent commit.

It also addresses one or maybe two of the comments that Ron
made. Instead of replacing the examples using piece, it presents
alternative overlay approaches to providing the same thing. Also while
addressing Tony's comments earlier, I may have addressed another one
of Ron's concerns. However, I will need to check that the current
language satisfies his concern.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Fix SIMD Lane example.

Markus pointed out that C/C++'s decay to a pointer semantics breaks my
example. However, Cary had a clever workaround that turned the pointer
into an implicit pointer which arguably it really is when some of its
values now exist in vector registers.

Bonus: a couple tiny touch-ups to text. One was a formatting error.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Fix registers used in example

I associated the wrong register with a variable and so they were
mismatched in the DWARF procedures.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Apply more of Cary's comments

* Reference the name of the figure rather than just the number for a
  figure to make Ron's job easier.

* Convert the 2nd implicit pointer example to have the overlays over
  an undefined.

* Adjust some of my notes for future work based upon some DWARF
  emitted by GCC in it's test suite.

* Adjust some directions on how to apply this submission.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Address Ron's concerns.

Most of the issues that Ron brought up were also brought up by Tony or
Cary in previous reviews of the issue. However, There was one place
where the text still needed to be changed.  This address that problem.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Make some requested modifications

There are 3 requested modifications included in this patch.

* Cary agreed that since overlays can concatenate pieces, it wasn't
  necessary to always have them on top of explicitly undefined
  storage.

* Pedro pointed out that my description of the second argument to
  overlay was confusing. I totally agree. It was obvious once he
  pointed it out.

* I didn't fully follow Tony's suggestions. I felt like I could
  tighten up the language slightly and address the first half of his
  comment cutting down the number of cases to be dealt with from 7 to
  4. I will need to discuss those with him and address them in a
  subsequent patch.

* Tony added some non-normative text at the beginning of section 3.12

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Revert a couple more examples

I missed a couple of examples that I wanted to revert to the simpler
form whcih doesn't set the overlays on undefined storage.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Add language back in which explains concatenation

I took out the language that explicitly said that a composite can
concatenate while I was working on Tony's comments. I didn't like that
part of my description anyway. After fussing with it for quite some
time I came up with something that I am much happier with. Now we can
all fight over it.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Minor touch ups and formatting fixes

Fixed a couple places where my formatting was broken and a few tiny
things in the text.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Update section numbers

Looking at the latest Locations on the stack proposal, I saw that the
section number for composites had been incremented. Update this
proposal to match that draft.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Address Baris's observation about length

Address Baris's concern that I was misleading readers in my
introductory examples about the length of composite storage when it is
placed over undefined storage. He was correct, the text seemed to
indicate that there was a finite length to the composite storage that
was created. Tried to fix the text.

Signed-off-by: Ben Woodard <woodard@redhat.com>

---------

Signed-off-by: Ben Woodard <woodard@redhat.com>