Change from undefined to composite (#110)
* Change from undefined to composite

Pedro had a good suggestion and to switch from basing examples on
undefined storage to basing it on empty composites. Made the changes
where it seemed like a good idea.

Signed-off-by: Ben Woodard <woodard@redhat.com>

* Add baris's changes

It was easier to apply Baris's changes by hand than it was to apply
his diff. Thus I did it this way.

Signed-off-by: Ben Woodard <woodard@redhat.com>

---------

Signed-off-by: Ben Woodard <woodard@redhat.com>