Update overlay description
Replace the description of the new overlay operator with a new
description largely written by Simon Marchi.

This was reviewed against https://github.com/barisaktemur/dwarf-locstack
and the text seems to address all the DW_OP_overlay tests with the
exception of the ones that have negative offsets. There continues to be
ongoing discussion about the value of negative offsets.

Simon's description of the overlay operator uses a novel way of
describing the operator. This is considered superior to the current
textually inline method of describing operators. Therefore all the other
DW_OP_operators are being converted to use this new format. The branch
where that is being done is
https://github.com/woodard/dwarf-locations/tree/op-formatting and the
actual proposal to move to the new format can be found in:
024-revise-operations.md

Signed-off-by: Ben Woodard <woodard@redhat.com>