Try out a new header format.
This reverts the changes of the name used for inline operands and
parameters. This was just too big of a change to make to the DWARF
standard and would have impacted a broad range of chapters not just
chapter 3. Without that change, the descriptions of the the operands
doesn't need to change and so the change ends up being just add a
header to each header. This makes this change much smaller.

With a header like this, many descriptions of the operators can be
greatly simplified. The sources of the information used by the
operators, their types as well as the output of the operator can be
removed from the operator description leaving the description to
describe the function of the operator.

Try out a new header format. Rather than a "bulky" set of lists inside
of list this compresses all of the information into a line.

Signed-off-by: Ben Woodard <woodard@redhat.com>