Address problmes that Markus pointed out.
Markus Metzger found 3 problems which pointed out another
problem. This patch addresses the issues that he pointed out.

When converting to the new convention where A is always on the left I
introduced a couple of bugs.
1) I had described drop incorrectly.
2) I had swapped arguments in the output of over.

I also had failed to paste in some boiler plate that I was using for
the desctiption of the changes to be made to section 3.2

His suggested change to swap also pointed out the need to explain that
generic stack argument names referred to the data and its type rather
than giving the stack position a name.

Signed-off-by: Ben Woodard <woodard@redhat.com>