Remove complaint about piece not composing
Early on when drafting this proposal, a big problem with the piece
operators was that they apparently didn't compose. This was due to
some of the side effects of moving them from being created on the side
of the stack rather than being created on the stack. Later on we
addressed those problems in the locations on the stack proposal when
we added the backward compatibility rules. This made it so that it was
more obvious that the problem was with the piece operators rather than
with the kind of composite storage that they created. However, the
text in the overlay proposal which pointed out the problem was never
edited to reflect this change in the locations on the stack
proposal. This removes most ot the text and edits the remaining text
to just show how overlay can be used.

Signed-off-by: Ben Woodard <woodard@redhat.com>