Present a proposed fix to refined type problem
This is a novel approach to fixing the refined type problem. Basically
rather than changing the behavior of location lists, or adding a
secondarly location list which specifies which provides the type for a
location, this addresses at least the most critical problems by
removing a restriction on DW_OP_reinterpret.

Signed-off-by: Ben Woodard <woodard@redhat.com>